# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Contacts API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/contacts": {
      "get": {
        "summary": "Get all contacts",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Contact" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new contact",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Contact" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Contact" }
              }
            }
          }
        }
      }
    },
    "/api/contacts/{id}": {
      "get": {
        "summary": "Get a contact by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Contact" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a contact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Contact" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a contact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": {
            "description": "Contact deleted"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Contact": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string" },
          "favoriteColor": { "type": "string" },
          "birthday": { "type": "string", "format": "date" }
        },
        "required": ["firstName", "lastName", "email", "favoriteColor", "birthday"]
      }
    }
  }
}